version updates

+++ app/src/main/java/com/gregor/lauritz/healthdashboard/domain/scoring/RhrBaselineProvider.kt
@@ -59,7 +59,7 @@ class AdaptiveRhrBaselineProvider
             return if (hasEnoughData) {
                 Math.round(baselineComputer.resolveBaselineRhrBpm(rhrValues, null))
             } else {
-                PhysiologyConstants.DEFAULT_RHR_BPM.toInt()
+                PhysiologyConstants.DEFAULT_RHR_BPM
             }
         }
